make relationship many to many post and tag

+++ app/Http/Controllers/CategoriesController.php
@@ -95,6 +95,13 @@ class CategoriesController extends Controller
      */
     public function destroy(Category $category)
     {
+        if($category->posts->count() > 0 ){
+
+            session()->flash('error', 'Category cannot delete becouse assosiated with a post');
+
+            return redirect()->back();
+        }
+
         $category->delete();
 
         session()->flash('success', 'Category delete successfully');
