+++ app/Http/Controllers/TagsController.php
@@ -95,6 +95,14 @@ class TagsController extends Controller
      */
     public function destroy(Tag $tag)
     {
+
+        if($tag->posts->count() > 0 ){
+
+            session()->flash('error', 'Tag cannot delete becouse assosiated with a post');
+
+            return redirect()->back();
+        }
+
         $tag->delete();
 
         session()->flash('success', 'Tag delete successfully');
